Laid groundwork for settings' Placeholder text
updated functions surrounding placeholder text.  Rearranged two of those functions because we need to announce a few variables at the offset.. Move them instead back to the top,  along with the needed variable declarations to the GLOBAL spot... and call it a day.  ALSO will need to figure out how to fix .insert (for placeholders)  to prevent "bad index text..