Minor tweak - Model questions
Trying to decide on where & how best to store intermittent data and how best to use the Model in this MVC architecture.  Model seems to either add an extra unnecessary layer,  or not be necessary at all...  without passing Model into View,  we would need to keep these variables as class/instance variables,  not within the Model obj.