Fixed UI & is functional
UI is now able to be displayed with the updated code.  This was accomplished by updating the protocol classes in both View & Presenter,  removed function creating the checkboxes---we will assess this at a later time---and by adjusting the @properties to avoid any name collisions by prefixing '_' to the attribute names.